+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/AbstractNetworkXmlContext.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.iidm.xml;
 
-import com.powsybl.iidm.anonymizer.Anonymizer;
+import com.powsybl.iidm.xml.anonymizer.Anonymizer;
 
 import java.util.Objects;
